Estimate network direct and indirect effects via a spectral method

TODO: Allow embedding L_tau(A) and L(A) in addition to A

Usage

netmediate(graph, formula, rank)

Arguments

graph

A tidygraph::tbl_graph() object.

formula

Details about the nodal design matrix. Of the form outcome ~ nodal_formula. For now, no interactiosn or fancy stuff are allowed in the formula.

rank

Integer describing desired rank of embedding dimension.

Value

A network_mediation object.

Examples

library(tidygraph)
#> 
#> Attaching package: ‘tidygraph’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter

smoking |>
  mutate(
    smokes_int = as.integer(smokes) - 1
  ) |>
  netmediate(smokes_int ~ sex + 0, rank = 7)
#> $formula
#> smokes_int ~ sex + 0
#> <environment: 0x55cc9e831df0>
#> 
#> $s
#> $s$d
#> [1] 34.974046 25.821247 23.674710 22.250565 19.014699 14.978582  9.522741
#> 
#> $s$u
#>              [,1]          [,2]          [,3]         [,4]         [,5]
#>   [1,] 0.06262231  0.1951661946 -0.0137040363  0.018374190 -0.034394363
#>   [2,] 0.05684293  0.1858050633 -0.0133772861  0.019296023 -0.035920163
#>   [3,] 0.04357454  0.1296855926 -0.0265796951  0.021609688 -0.016137885
#>   [4,] 0.05382365  0.1651411072 -0.0314199506  0.018308537 -0.015954231
#>   [5,] 0.03551680  0.1114009734 -0.0081342365  0.030097213 -0.020989556
#>   [6,] 0.04103728  0.1455052653 -0.0173691749  0.007838749 -0.030307498
#>   [7,] 0.04981072  0.1939557268 -0.0236352480  0.032732946 -0.028497606
#>   [8,] 0.04403088  0.1365052829 -0.0240301300  0.027109327 -0.031221360
#>   [9,] 0.04123137  0.1616962822 -0.0240621081  0.021712229 -0.018465547
#>  [10,] 0.03077918  0.1176495625 -0.0232788594  0.021091630 -0.022604727
#>  [11,] 0.05224733  0.1943666897 -0.0283804835  0.024658450 -0.022040731
#>  [12,] 0.05797059  0.1685437423 -0.0318761678  0.029725018 -0.029797205
#>  [13,] 0.06108193  0.1945889796 -0.0239164908  0.015535025 -0.005190434
#>  [14,] 0.04547123  0.1653560103 -0.0285214358  0.024718776 -0.023533233
#>  [15,] 0.04856696  0.1943781818 -0.0285824221  0.027099531 -0.026173516
#>  [16,] 0.05479749  0.1910635718 -0.0360551701  0.031348366 -0.029649764
#>  [17,] 0.06066330  0.1443080550 -0.0204042654  0.013428320 -0.014853443
#>  [18,] 0.06117021  0.1980574214 -0.0252133821  0.015756523  0.005444599
#>  [19,] 0.05347635  0.1948180381 -0.0213868133  0.025177595 -0.034329352
#>  [20,] 0.05437063  0.1253552575 -0.0278079176  0.012963896  0.007864431
#>  [21,] 0.03594577  0.1289507908 -0.0182500693  0.014443081  0.001327309
#>  [22,] 0.02620956  0.0964887314 -0.0169306836  0.015269217 -0.018681247
#>  [23,] 0.10692770  0.1384945320  0.0003277783  0.009504068  0.037386795
#>  [24,] 0.10971915  0.1851565741 -0.0193838477  0.016381613  0.004023934
#>  [25,] 0.05219354  0.1513704716 -0.0185270390  0.009362188  0.028064447
#>  [26,] 0.07002402  0.1721759363 -0.0110036950  0.012856063 -0.011006542
#>  [27,] 0.04856696  0.1943781818 -0.0285824221  0.027099531 -0.026173516
#>  [28,] 0.03855196  0.1258160544 -0.0035394439  0.022879073 -0.023443762
#>  [29,] 0.03113608  0.1340479228 -0.0204351899  0.018902058 -0.021666880
#>  [30,] 0.06022161  0.1576183544 -0.0287288097  0.018971592 -0.028029405
#>  [31,] 0.07478379  0.1401252325 -0.0260513621  0.005062996  0.007320206
#>  [32,] 0.04822163  0.1500231222 -0.0157053445  0.006472252 -0.022108004
#>  [33,] 0.03556265  0.1254723110 -0.0207175751  0.014293633 -0.024116802
#>  [34,] 0.08133132 -0.0243175625  0.0483247039 -0.145079122 -0.102013713
#>  [35,] 0.13118293 -0.0018277019  0.0535030265 -0.151573243 -0.034885773
#>  [36,] 0.07743930 -0.0274720999  0.0328169078 -0.139532314 -0.121267228
#>  [37,] 0.09699700  0.0086742352  0.0326846871 -0.131011683 -0.115647112
#>  [38,] 0.09248950 -0.0055248773  0.0393296409 -0.124724996 -0.058195104
#>  [39,] 0.09296627 -0.0175118096  0.0597297598 -0.159877319 -0.079302346
#>  [40,] 0.07486804 -0.0221072675  0.0527540233 -0.145336446 -0.105849070
#>  [41,] 0.09916568 -0.0059531747  0.0551919692 -0.120848137 -0.123000592
#>  [42,] 0.14637902 -0.0374036897  0.0389715467 -0.132904279 -0.037212532
#>  [43,] 0.08558374 -0.0213565197  0.0640437678 -0.114663301 -0.069608919
#>  [44,] 0.11239853 -0.0395976615  0.0276325206 -0.145951129 -0.072113056
#>  [45,] 0.06829548 -0.0253785112  0.0234595971 -0.123898729 -0.112226502
#>  [46,] 0.09165731 -0.0080106231  0.0448666426 -0.142198999 -0.090400275
#>  [47,] 0.05825542 -0.0178368834  0.0356794687 -0.124446312 -0.084134243
#>  [48,] 0.06179709 -0.0194460503  0.0328733109 -0.121395878 -0.083177201
#>  [49,] 0.09180466 -0.0276748649  0.0400585808 -0.152870259 -0.082390596
#>  [50,] 0.08586964 -0.0244450388  0.0925576308 -0.125715645 -0.100267533
#>  [51,] 0.10675664 -0.0010188410  0.0667037467 -0.150555909 -0.074689370
#>  [52,] 0.08829449 -0.0230945536  0.0521656211 -0.152885329 -0.090121049
#>  [53,] 0.06964399 -0.0204252076  0.0522431358 -0.154239674 -0.099510355
#>  [54,] 0.05280587 -0.0135342888  0.0389849351 -0.118154080 -0.078914985
#>  [55,] 0.06625157 -0.0232272866  0.0370121699 -0.100644955 -0.087268819
#>  [56,] 0.07507634 -0.0236301611  0.0451548226 -0.147554250 -0.105561419
#>  [57,] 0.13569695 -0.0189632306  0.0633401782 -0.135536368 -0.054056323
#>  [58,] 0.05749903 -0.0122766148  0.0352416989 -0.117996298 -0.086179470
#>  [59,] 0.10136522 -0.0408702715  0.0483861787 -0.070065264 -0.093413621
#>  [60,] 0.05017023 -0.0180092045  0.0038250468 -0.021529914  0.090410150
#>  [61,] 0.11733765 -0.0323095789  0.0543873277 -0.005272854  0.189218581
#>  [62,] 0.06298203  0.0129091588  0.0136555729 -0.020116682  0.144023152
#>  [63,] 0.04622766 -0.0120144366  0.0171948638 -0.027246462  0.139013363
#>  [64,] 0.04618866 -0.0062104600  0.0250971771 -0.017583101  0.176765920
#>  [65,] 0.04747053 -0.0175084652  0.0228539616  0.013385100  0.122236176
#>  [66,] 0.03118825 -0.0064380936  0.0159308144 -0.012233537  0.120144422
#>  [67,] 0.03611621 -0.0100311669  0.0149475026 -0.007228915  0.129738513
#>  [68,] 0.04707034 -0.0071564233  0.0203990692 -0.025789381  0.174897555
#>  [69,] 0.10821809 -0.0131009158  0.0040690655 -0.015136713  0.205738353
#>  [70,] 0.05801351 -0.0094609779  0.0382000126 -0.019502109  0.197868456
#>  [71,] 0.04418828 -0.0121222546  0.0235169749 -0.009963756  0.106039843
#>  [72,] 0.09812716 -0.0281722585  0.0073404962 -0.007295424  0.213197500
#>  [73,] 0.06186651 -0.0068755896  0.0120845548 -0.022482243  0.142706004
#>  [74,] 0.03585346 -0.0085058727  0.0075163511 -0.015233560  0.133971058
#>  [75,] 0.13717025 -0.0119300798  0.0458402462 -0.019365604  0.111236393
#>  [76,] 0.12076275 -0.0107569314  0.0557810887 -0.084021094  0.196605630
#>  [77,] 0.08628248  0.0442461352  0.0066913704 -0.035049388  0.167837982
#>  [78,] 0.06400341 -0.0103880012  0.0214780891 -0.026126305  0.132789134
#>  [79,] 0.03603610 -0.0048774009  0.0077076213 -0.014285644  0.137038743
#>  [80,] 0.03985373 -0.0061391643  0.0111307682 -0.033348326  0.108813480
#>  [81,] 0.05943502 -0.0122045038  0.0223614368 -0.022936110  0.187300354
#>  [82,] 0.04587544 -0.0102533581  0.0282470116 -0.020751214  0.093910839
#>  [83,] 0.12092810  0.0223466256  0.0024961369 -0.036880106  0.210057012
#>  [84,] 0.10045841  0.0032479346  0.0219274779 -0.070436578  0.189076948
#>  [85,] 0.05201445  0.0210075743  0.0067839786 -0.020111922  0.145794683
#>  [86,] 0.04498886 -0.0156581370  0.0129134292  0.002248998  0.102083691
#>  [87,] 0.09066068 -0.0662236798 -0.1191099164  0.068568973 -0.022060906
#>  [88,] 0.10435099 -0.0716020286 -0.1170304600  0.059904962 -0.029230195
#>  [89,] 0.09591013 -0.0586665002 -0.1265929457  0.076736665 -0.059339329
#>  [90,] 0.09046599 -0.0702781137 -0.1263034001  0.084064315 -0.055108373
#>  [91,] 0.09215900 -0.0626045111 -0.1341390402  0.067975486 -0.057675547
#>  [92,] 0.11763425 -0.0744641404 -0.1156974198  0.064250233 -0.022053211
#>  [93,] 0.15113558 -0.0565860409 -0.1024768746  0.042658305  0.029156915
#>  [94,] 0.11819276 -0.0448331299 -0.1174495004  0.069682029 -0.066297846
#>  [95,] 0.13561514 -0.0395911029 -0.0661685630  0.027323634 -0.050016378
#>  [96,] 0.11022075 -0.0637827894 -0.1216617635  0.064579814 -0.039371388
#>  [97,] 0.12359925 -0.0493147924 -0.1160455632  0.052278756  0.004932938
#>  [98,] 0.09520758 -0.0661336327 -0.1333920165  0.070723422 -0.022579506
#>  [99,] 0.08852228 -0.0493108719 -0.1182193256  0.071464183 -0.019559688
#> [100,] 0.10199517 -0.0625023076 -0.1092109900  0.078632485 -0.019433487
#> [101,] 0.10949522 -0.0659966596 -0.1184514977  0.055057591 -0.040810802
#> [102,] 0.07449925 -0.0532786119 -0.0838155532  0.052445599 -0.030850366
#> [103,] 0.11054376 -0.0651286187 -0.0848133019  0.041644003 -0.021783950
#> [104,] 0.16776013 -0.0579919276 -0.0947890032  0.015966514  0.026373663
#> [105,] 0.08929843 -0.0676509221 -0.1335008433  0.077888442 -0.041098241
#> [106,] 0.09358811 -0.0689563977 -0.1290391373  0.069596182 -0.044948746
#> [107,] 0.09630093 -0.0668935732 -0.1291282228  0.084596597 -0.022364143
#> [108,] 0.09949210 -0.0701651083 -0.1108244226  0.071647696 -0.055764357
#> [109,] 0.10603159 -0.0679124228 -0.1201655726  0.069113680 -0.001796769
#> [110,] 0.09147780 -0.0624283834 -0.1337433423  0.083266397 -0.039209360
#> [111,] 0.09167974 -0.0646647451 -0.1330415589  0.070872584 -0.058942056
#> [112,] 0.09077085 -0.0640629339 -0.1294243663  0.082507244 -0.055207238
#> [113,] 0.09378986 -0.0491869914 -0.1018761864  0.055922568 -0.016442252
#> [114,] 0.09771455 -0.0699881586 -0.1275230846  0.048746477 -0.066374050
#> [115,] 0.14714386 -0.0723529121 -0.0863009683  0.052201012 -0.035950659
#> [116,] 0.04738547 -0.0196070032  0.1561812562  0.138368817 -0.039032952
#> [117,] 0.02836372 -0.0120071081  0.0834333249  0.067021530 -0.014848270
#> [118,] 0.03387649 -0.0197966564  0.0785913009  0.097757399 -0.036155250
#> [119,] 0.04858488 -0.0195828842  0.1583776132  0.138408009 -0.043463248
#> [120,] 0.05295043 -0.0180185509  0.1444582704  0.107990424 -0.030500412
#> [121,] 0.06754172 -0.0182892105  0.1539850220  0.135793574 -0.027071151
#> [122,] 0.04859916 -0.0155838689  0.1332736780  0.092443071 -0.035841685
#> [123,] 0.05077267 -0.0197717208  0.1586949266  0.132510197 -0.047787212
#> [124,] 0.02847493 -0.0148588244  0.0801863593  0.080460897 -0.014633755
#> [125,] 0.03284169 -0.0131980927  0.1123161226  0.101830143 -0.033982861
#> [126,] 0.06531472 -0.0203211664  0.1703541492  0.117814780 -0.042757615
#> [127,] 0.05461326 -0.0166079372  0.1343961503  0.086333125 -0.014095737
#> [128,] 0.05331193 -0.0094484726  0.1246074493  0.101839576 -0.023943671
#> [129,] 0.04474296 -0.0113307718  0.1220700867  0.130980091 -0.046515964
#> [130,] 0.04938224 -0.0272979046  0.0676637571  0.091086602 -0.026899372
#> [131,] 0.11472602 -0.0304625887  0.1623155316  0.099993201  0.041255739
#> [132,] 0.05442077 -0.0115689215  0.1142688415  0.091884049 -0.004327310
#> [133,] 0.07257124 -0.0172428002  0.0881092549  0.066325456  0.003766827
#> [134,] 0.04346204 -0.0143067889  0.1263097455  0.102292110 -0.021747724
#> [135,] 0.05033496 -0.0205829864  0.1584434487  0.137432763 -0.023369830
#> [136,] 0.07744479  0.0177506686  0.1223575929  0.076162040 -0.040236020
#> [137,] 0.04393008 -0.0217277057  0.0964575903  0.096534804 -0.039603311
#> [138,] 0.03424794 -0.0092221856  0.0972205004  0.091990851 -0.033918697
#> [139,] 0.02788098 -0.0145096750  0.0805323041  0.081799550 -0.018629458
#> [140,] 0.04269385 -0.0169687194  0.1166209061  0.094299598 -0.045904049
#> [141,] 0.06964749 -0.0154983573  0.1643106298  0.120735377 -0.056018862
#> [142,] 0.09606941 -0.0074415351  0.1212719282  0.081176125 -0.016714625
#> [143,] 0.03611613 -0.0139515944  0.0987029929  0.060572411 -0.037840663
#> [144,] 0.04242339 -0.0185686815  0.1282847197  0.107985997 -0.027892346
#> [145,] 0.07491364 -0.0293482572  0.1409971040  0.149055892 -0.041209826
#> [146,] 0.07201938 -0.0308558356  0.1460212190  0.136025656 -0.043397177
#> [147,] 0.08813624 -0.0150112369  0.0671079000  0.003160240  0.013412656
#> [148,] 0.06767412 -0.0141228932  0.0197455467 -0.013119086  0.049870394
#> [149,] 0.04725252 -0.0122263941  0.0280846301 -0.042963579  0.016498276
#> [150,] 0.05712851 -0.0233146103 -0.0106265192 -0.015342507  0.028312222
#> [151,] 0.09369239 -0.0062721267  0.0348605360 -0.018985060  0.067761192
#> [152,] 0.12835980  0.0031709559  0.0241019009 -0.078639665  0.082749344
#> [153,] 0.13239762 -0.0003082346  0.0049827778 -0.056756511  0.082716459
#> [154,] 0.07276650 -0.0123644559  0.0171313599 -0.035674128  0.047012484
#> [155,] 0.05815623 -0.0181548358  0.0147414003 -0.041082249  0.016725725
#> [156,] 0.10240034 -0.0058468654  0.0257310203 -0.058684280  0.080334255
#> [157,] 0.06478343 -0.0178670164 -0.0043098246 -0.016318394  0.030764226
#> [158,] 0.03813960 -0.0107468437  0.0087575733 -0.024700461  0.040717537
#> [159,] 0.03669184 -0.0093823332  0.0084545724 -0.023938428  0.026649079
#> [160,] 0.05491259 -0.0178222021  0.0010990852 -0.014022226  0.046987371
#> [161,] 0.03866114 -0.0110132470  0.0101864876 -0.029517651  0.029213018
#> [162,] 0.04271390 -0.0043239374  0.0079831041 -0.021221001  0.057071532
#>                 [,6]          [,7]
#>   [1,] -0.0003161633 -0.0147973992
#>   [2,]  0.0102913629 -0.0072249754
#>   [3,] -0.0106281430 -0.0192321006
#>   [4,]  0.0103817881  0.0025755801
#>   [5,]  0.0018599033 -0.0007530624
#>   [6,]  0.0150681482  0.0246015833
#>   [7,]  0.0126900819  0.0348266024
#>   [8,]  0.0086381275 -0.0007310646
#>   [9,]  0.0093641942  0.0216417827
#>  [10,]  0.0189037436  0.0509418314
#>  [11,] -0.0011286673  0.0218479375
#>  [12,]  0.0119440064  0.0353348469
#>  [13,]  0.0259094013  0.0119124660
#>  [14,]  0.0019145582  0.0067206625
#>  [15,]  0.0112068679  0.0283712813
#>  [16,]  0.0173622938  0.0530857137
#>  [17,]  0.0075555312  0.0291821672
#>  [18,]  0.0160876767  0.0035570374
#>  [19,]  0.0151178617  0.0217990235
#>  [20,]  0.0443193450  0.0571444397
#>  [21,]  0.0285862769  0.0415508841
#>  [22,]  0.0191293099  0.0493139808
#>  [23,] -0.1380124178 -0.1283069899
#>  [24,]  0.0102008754 -0.0363641972
#>  [25,]  0.0045769044 -0.0237280957
#>  [26,]  0.0010725818 -0.0665145782
#>  [27,]  0.0112068679  0.0283712813
#>  [28,] -0.0003488678 -0.0290969034
#>  [29,]  0.0226419922  0.0562085795
#>  [30,]  0.0366512777  0.0723216070
#>  [31,] -0.0509366006 -0.0649614716
#>  [32,] -0.0145407808 -0.0049625576
#>  [33,]  0.0223005171  0.0439077350
#>  [34,]  0.0357222921  0.0836061183
#>  [35,] -0.0584661301 -0.0541668829
#>  [36,]  0.0719759900  0.0545613804
#>  [37,]  0.0508095892  0.0460471442
#>  [38,]  0.0103632369 -0.0984933066
#>  [39,] -0.0189575668  0.0038073649
#>  [40,]  0.0451810784  0.0470584198
#>  [41,]  0.0651187141  0.0964619213
#>  [42,]  0.0561698336 -0.1144410531
#>  [43,]  0.0393066015 -0.1089504144
#>  [44,]  0.0488794979 -0.0513525747
#>  [45,]  0.0656496741  0.0276955250
#>  [46,]  0.0532183510  0.0833043076
#>  [47,]  0.0678712541  0.0472163002
#>  [48,]  0.0543758550  0.0594135973
#>  [49,]  0.0714186610 -0.0288995707
#>  [50,]  0.0575679134 -0.0273191816
#>  [51,] -0.0203042458 -0.0544854900
#>  [52,]  0.0358693711 -0.0716046628
#>  [53,]  0.0321287180  0.0828383063
#>  [54,]  0.0152187668  0.0331915953
#>  [55,]  0.0454984693  0.0982177853
#>  [56,]  0.0646643672  0.0723387098
#>  [57,] -0.0006692876 -0.0986438971
#>  [58,]  0.0591171105  0.0444723518
#>  [59,]  0.0116238722  0.0685833708
#>  [60,]  0.1068981129  0.1413466975
#>  [61,]  0.0897957690 -0.0643792865
#>  [62,]  0.0390047665 -0.0825127729
#>  [63,]  0.1039486040  0.0870243645
#>  [64,]  0.1108129066  0.0644192491
#>  [65,]  0.1045400313  0.1244765212
#>  [66,]  0.0540143109  0.0847501878
#>  [67,]  0.0890897359  0.1068962137
#>  [68,]  0.1108621395  0.0658717294
#>  [69,]  0.0772773583  0.1073083590
#>  [70,]  0.1241603480  0.0609842047
#>  [71,]  0.0967972854  0.1216687142
#>  [72,]  0.0679319810  0.0090484939
#>  [73,]  0.0562377001 -0.0417906950
#>  [74,]  0.0716112530  0.0869336639
#>  [75,]  0.0466658189 -0.2447689615
#>  [76,] -0.0119113732 -0.0918778945
#>  [77,]  0.0127864805 -0.0891119095
#>  [78,]  0.0336157137 -0.1268198677
#>  [79,]  0.0771419728  0.0341251437
#>  [80,]  0.0789325167  0.0325924600
#>  [81,]  0.1181567011  0.1232904522
#>  [82,]  0.0612350172  0.0170767550
#>  [83,]  0.0367043776  0.0023067134
#>  [84,]  0.0983370831 -0.0140338903
#>  [85,]  0.0479589070 -0.0376289923
#>  [86,]  0.0300604895  0.1478645673
#>  [87,]  0.0019957624  0.1136191319
#>  [88,]  0.0530228117 -0.0338389396
#>  [89,]  0.0362986201  0.0205656927
#>  [90,]  0.0265387625  0.0161493907
#>  [91,]  0.0261726424 -0.0045500326
#>  [92,] -0.0260891185 -0.0366203793
#>  [93,] -0.0722373926 -0.0867031199
#>  [94,]  0.0181171171  0.0195658378
#>  [95,]  0.0797044585  0.1447474533
#>  [96,] -0.0522517788 -0.0342001741
#>  [97,] -0.0127486729 -0.0970627310
#>  [98,]  0.0192588588 -0.0139303737
#>  [99,]  0.0186505157  0.1153161048
#> [100,]  0.0018980044 -0.0503097542
#> [101,]  0.0406286094  0.0687818583
#> [102,] -0.0598615490  0.0010364633
#> [103,]  0.0298055335 -0.1621188245
#> [104,] -0.0710480970 -0.1827067202
#> [105,]  0.0249925099 -0.0286454817
#> [106,]  0.0392831792  0.0130126833
#> [107,]  0.0328944151  0.0734893743
#> [108,] -0.0073730775  0.0158722663
#> [109,]  0.0331901657 -0.0602092146
#> [110,]  0.0289702339  0.0208707048
#> [111,]  0.0278445885  0.0062487218
#> [112,]  0.0248232731  0.0075189508
#> [113,]  0.0043844451 -0.0426877743
#> [114,]  0.0432696302  0.0262288267
#> [115,]  0.0425445915  0.1139000496
#> [116,]  0.0182342348  0.0247095787
#> [117,]  0.0285455654  0.0703907550
#> [118,]  0.0229959728  0.0710407119
#> [119,]  0.0011910345  0.0109654657
#> [120,] -0.0294583574 -0.0365527163
#> [121,]  0.0410297916  0.0278491691
#> [122,] -0.0010536127 -0.0679482313
#> [123,]  0.0057779465  0.0256744395
#> [124,]  0.0307022293  0.0747034641
#> [125,]  0.0163051326  0.0259224064
#> [126,]  0.0227340867 -0.0163809978
#> [127,] -0.0114040140 -0.0774713521
#> [128,]  0.0071966560 -0.0691489768
#> [129,]  0.0236996551  0.0679276729
#> [130,]  0.0210028412  0.1336338167
#> [131,] -0.0582407606 -0.0720738632
#> [132,]  0.0080637500 -0.0705758696
#> [133,] -0.0394293338 -0.0820956898
#> [134,] -0.0072447828 -0.0716512815
#> [135,]  0.0363480784  0.0457659869
#> [136,] -0.0026265807 -0.1656200792
#> [137,]  0.0367528980  0.0869717675
#> [138,]  0.0104643900  0.0931635013
#> [139,]  0.0213804045  0.0629510938
#> [140,]  0.0170692917  0.0359846153
#> [141,]  0.0108371993  0.0532992988
#> [142,] -0.0320092033 -0.2019958681
#> [143,] -0.0040800718 -0.0238096700
#> [144,]  0.0233066278  0.0478770087
#> [145,]  0.0166861053  0.0287040837
#> [146,]  0.0265195470  0.0799376812
#> [147,] -0.2422656791  0.0032388327
#> [148,] -0.2340920451  0.0615556298
#> [149,] -0.1914025000  0.1624804733
#> [150,] -0.1817309986  0.2143912183
#> [151,] -0.2322600088  0.0156263112
#> [152,] -0.1799912550 -0.0781985192
#> [153,] -0.1971043577 -0.0686434524
#> [154,] -0.2371493403  0.0226995543
#> [155,] -0.1725165109  0.2011126015
#> [156,] -0.1960483170  0.0130893057
#> [157,] -0.2306799322  0.0815554471
#> [158,] -0.1969791232  0.1586071112
#> [159,] -0.1861615176  0.1292591070
#> [160,] -0.1917600418  0.1018367716
#> [161,] -0.2002183317  0.1443453076
#> [162,] -0.1988799064  0.0761857859
#> 
#> $s$v
#>              [,1]          [,2]          [,3]         [,4]         [,5]
#>   [1,] 0.06262231  0.1951661946 -0.0137040363  0.018374190 -0.034394363
#>   [2,] 0.05684293  0.1858050633 -0.0133772861  0.019296023 -0.035920163
#>   [3,] 0.04357454  0.1296855926 -0.0265796951  0.021609688 -0.016137885
#>   [4,] 0.05382365  0.1651411072 -0.0314199506  0.018308537 -0.015954231
#>   [5,] 0.03551680  0.1114009734 -0.0081342365  0.030097213 -0.020989556
#>   [6,] 0.04103728  0.1455052653 -0.0173691749  0.007838749 -0.030307498
#>   [7,] 0.04981072  0.1939557268 -0.0236352480  0.032732946 -0.028497606
#>   [8,] 0.04403088  0.1365052829 -0.0240301300  0.027109327 -0.031221360
#>   [9,] 0.04123137  0.1616962822 -0.0240621081  0.021712229 -0.018465547
#>  [10,] 0.03077918  0.1176495625 -0.0232788594  0.021091630 -0.022604727
#>  [11,] 0.05224733  0.1943666897 -0.0283804835  0.024658450 -0.022040731
#>  [12,] 0.05797059  0.1685437423 -0.0318761678  0.029725018 -0.029797205
#>  [13,] 0.06108193  0.1945889796 -0.0239164908  0.015535025 -0.005190434
#>  [14,] 0.04547123  0.1653560103 -0.0285214358  0.024718776 -0.023533233
#>  [15,] 0.04856696  0.1943781818 -0.0285824221  0.027099531 -0.026173516
#>  [16,] 0.05479749  0.1910635718 -0.0360551701  0.031348366 -0.029649764
#>  [17,] 0.06066330  0.1443080550 -0.0204042654  0.013428320 -0.014853443
#>  [18,] 0.06117021  0.1980574214 -0.0252133821  0.015756523  0.005444599
#>  [19,] 0.05347635  0.1948180381 -0.0213868133  0.025177595 -0.034329352
#>  [20,] 0.05437063  0.1253552575 -0.0278079176  0.012963896  0.007864431
#>  [21,] 0.03594577  0.1289507908 -0.0182500693  0.014443081  0.001327309
#>  [22,] 0.02620956  0.0964887314 -0.0169306836  0.015269217 -0.018681247
#>  [23,] 0.10692770  0.1384945320  0.0003277783  0.009504068  0.037386795
#>  [24,] 0.10971915  0.1851565741 -0.0193838477  0.016381613  0.004023934
#>  [25,] 0.05219354  0.1513704716 -0.0185270390  0.009362188  0.028064447
#>  [26,] 0.07002402  0.1721759363 -0.0110036950  0.012856063 -0.011006542
#>  [27,] 0.04856696  0.1943781818 -0.0285824221  0.027099531 -0.026173516
#>  [28,] 0.03855196  0.1258160544 -0.0035394439  0.022879073 -0.023443762
#>  [29,] 0.03113608  0.1340479228 -0.0204351899  0.018902058 -0.021666880
#>  [30,] 0.06022161  0.1576183544 -0.0287288097  0.018971592 -0.028029405
#>  [31,] 0.07478379  0.1401252325 -0.0260513621  0.005062996  0.007320206
#>  [32,] 0.04822163  0.1500231222 -0.0157053445  0.006472252 -0.022108004
#>  [33,] 0.03556265  0.1254723110 -0.0207175751  0.014293633 -0.024116802
#>  [34,] 0.08133132 -0.0243175625  0.0483247039 -0.145079122 -0.102013713
#>  [35,] 0.13118293 -0.0018277019  0.0535030265 -0.151573243 -0.034885773
#>  [36,] 0.07743930 -0.0274720999  0.0328169078 -0.139532314 -0.121267228
#>  [37,] 0.09699700  0.0086742352  0.0326846871 -0.131011683 -0.115647112
#>  [38,] 0.09248950 -0.0055248773  0.0393296409 -0.124724996 -0.058195104
#>  [39,] 0.09296627 -0.0175118096  0.0597297598 -0.159877319 -0.079302346
#>  [40,] 0.07486804 -0.0221072675  0.0527540233 -0.145336446 -0.105849070
#>  [41,] 0.09916568 -0.0059531747  0.0551919692 -0.120848137 -0.123000592
#>  [42,] 0.14637902 -0.0374036897  0.0389715467 -0.132904279 -0.037212532
#>  [43,] 0.08558374 -0.0213565197  0.0640437678 -0.114663301 -0.069608919
#>  [44,] 0.11239853 -0.0395976615  0.0276325206 -0.145951129 -0.072113056
#>  [45,] 0.06829548 -0.0253785112  0.0234595971 -0.123898729 -0.112226502
#>  [46,] 0.09165731 -0.0080106231  0.0448666426 -0.142198999 -0.090400275
#>  [47,] 0.05825542 -0.0178368834  0.0356794687 -0.124446312 -0.084134243
#>  [48,] 0.06179709 -0.0194460503  0.0328733109 -0.121395878 -0.083177201
#>  [49,] 0.09180466 -0.0276748649  0.0400585808 -0.152870259 -0.082390596
#>  [50,] 0.08586964 -0.0244450388  0.0925576308 -0.125715645 -0.100267533
#>  [51,] 0.10675664 -0.0010188410  0.0667037467 -0.150555909 -0.074689370
#>  [52,] 0.08829449 -0.0230945536  0.0521656211 -0.152885329 -0.090121049
#>  [53,] 0.06964399 -0.0204252076  0.0522431358 -0.154239674 -0.099510355
#>  [54,] 0.05280587 -0.0135342888  0.0389849351 -0.118154080 -0.078914985
#>  [55,] 0.06625157 -0.0232272866  0.0370121699 -0.100644955 -0.087268819
#>  [56,] 0.07507634 -0.0236301611  0.0451548226 -0.147554250 -0.105561419
#>  [57,] 0.13569695 -0.0189632306  0.0633401782 -0.135536368 -0.054056323
#>  [58,] 0.05749903 -0.0122766148  0.0352416989 -0.117996298 -0.086179470
#>  [59,] 0.10136522 -0.0408702715  0.0483861787 -0.070065264 -0.093413621
#>  [60,] 0.05017023 -0.0180092045  0.0038250468 -0.021529914  0.090410150
#>  [61,] 0.11733765 -0.0323095789  0.0543873277 -0.005272854  0.189218581
#>  [62,] 0.06298203  0.0129091588  0.0136555729 -0.020116682  0.144023152
#>  [63,] 0.04622766 -0.0120144366  0.0171948638 -0.027246462  0.139013363
#>  [64,] 0.04618866 -0.0062104600  0.0250971771 -0.017583101  0.176765920
#>  [65,] 0.04747053 -0.0175084652  0.0228539616  0.013385100  0.122236176
#>  [66,] 0.03118825 -0.0064380936  0.0159308144 -0.012233537  0.120144422
#>  [67,] 0.03611621 -0.0100311669  0.0149475026 -0.007228915  0.129738513
#>  [68,] 0.04707034 -0.0071564233  0.0203990692 -0.025789381  0.174897555
#>  [69,] 0.10821809 -0.0131009158  0.0040690655 -0.015136713  0.205738353
#>  [70,] 0.05801351 -0.0094609779  0.0382000126 -0.019502109  0.197868456
#>  [71,] 0.04418828 -0.0121222546  0.0235169749 -0.009963756  0.106039843
#>  [72,] 0.09812716 -0.0281722585  0.0073404962 -0.007295424  0.213197500
#>  [73,] 0.06186651 -0.0068755896  0.0120845548 -0.022482243  0.142706004
#>  [74,] 0.03585346 -0.0085058727  0.0075163511 -0.015233560  0.133971058
#>  [75,] 0.13717025 -0.0119300798  0.0458402462 -0.019365604  0.111236393
#>  [76,] 0.12076275 -0.0107569314  0.0557810887 -0.084021094  0.196605630
#>  [77,] 0.08628248  0.0442461352  0.0066913704 -0.035049388  0.167837982
#>  [78,] 0.06400341 -0.0103880012  0.0214780891 -0.026126305  0.132789134
#>  [79,] 0.03603610 -0.0048774009  0.0077076213 -0.014285644  0.137038743
#>  [80,] 0.03985373 -0.0061391643  0.0111307682 -0.033348326  0.108813480
#>  [81,] 0.05943502 -0.0122045038  0.0223614368 -0.022936110  0.187300354
#>  [82,] 0.04587544 -0.0102533581  0.0282470116 -0.020751214  0.093910839
#>  [83,] 0.12092810  0.0223466256  0.0024961369 -0.036880106  0.210057012
#>  [84,] 0.10045841  0.0032479346  0.0219274779 -0.070436578  0.189076948
#>  [85,] 0.05201445  0.0210075743  0.0067839786 -0.020111922  0.145794683
#>  [86,] 0.04498886 -0.0156581370  0.0129134292  0.002248998  0.102083691
#>  [87,] 0.09066068 -0.0662236798 -0.1191099164  0.068568973 -0.022060906
#>  [88,] 0.10435099 -0.0716020286 -0.1170304600  0.059904962 -0.029230195
#>  [89,] 0.09591013 -0.0586665002 -0.1265929457  0.076736665 -0.059339329
#>  [90,] 0.09046599 -0.0702781137 -0.1263034001  0.084064315 -0.055108373
#>  [91,] 0.09215900 -0.0626045111 -0.1341390402  0.067975486 -0.057675547
#>  [92,] 0.11763425 -0.0744641404 -0.1156974198  0.064250233 -0.022053211
#>  [93,] 0.15113558 -0.0565860409 -0.1024768746  0.042658305  0.029156915
#>  [94,] 0.11819276 -0.0448331299 -0.1174495004  0.069682029 -0.066297846
#>  [95,] 0.13561514 -0.0395911029 -0.0661685630  0.027323634 -0.050016378
#>  [96,] 0.11022075 -0.0637827894 -0.1216617635  0.064579814 -0.039371388
#>  [97,] 0.12359925 -0.0493147924 -0.1160455632  0.052278756  0.004932938
#>  [98,] 0.09520758 -0.0661336327 -0.1333920165  0.070723422 -0.022579506
#>  [99,] 0.08852228 -0.0493108719 -0.1182193256  0.071464183 -0.019559688
#> [100,] 0.10199517 -0.0625023076 -0.1092109900  0.078632485 -0.019433487
#> [101,] 0.10949522 -0.0659966596 -0.1184514977  0.055057591 -0.040810802
#> [102,] 0.07449925 -0.0532786119 -0.0838155532  0.052445599 -0.030850366
#> [103,] 0.11054376 -0.0651286187 -0.0848133019  0.041644003 -0.021783950
#> [104,] 0.16776013 -0.0579919276 -0.0947890032  0.015966514  0.026373663
#> [105,] 0.08929843 -0.0676509221 -0.1335008433  0.077888442 -0.041098241
#> [106,] 0.09358811 -0.0689563977 -0.1290391373  0.069596182 -0.044948746
#> [107,] 0.09630093 -0.0668935732 -0.1291282228  0.084596597 -0.022364143
#> [108,] 0.09949210 -0.0701651083 -0.1108244226  0.071647696 -0.055764357
#> [109,] 0.10603159 -0.0679124228 -0.1201655726  0.069113680 -0.001796769
#> [110,] 0.09147780 -0.0624283834 -0.1337433423  0.083266397 -0.039209360
#> [111,] 0.09167974 -0.0646647451 -0.1330415589  0.070872584 -0.058942056
#> [112,] 0.09077085 -0.0640629339 -0.1294243663  0.082507244 -0.055207238
#> [113,] 0.09378986 -0.0491869914 -0.1018761864  0.055922568 -0.016442252
#> [114,] 0.09771455 -0.0699881586 -0.1275230846  0.048746477 -0.066374050
#> [115,] 0.14714386 -0.0723529121 -0.0863009683  0.052201012 -0.035950659
#> [116,] 0.04738547 -0.0196070032  0.1561812562  0.138368817 -0.039032952
#> [117,] 0.02836372 -0.0120071081  0.0834333249  0.067021530 -0.014848270
#> [118,] 0.03387649 -0.0197966564  0.0785913009  0.097757399 -0.036155250
#> [119,] 0.04858488 -0.0195828842  0.1583776132  0.138408009 -0.043463248
#> [120,] 0.05295043 -0.0180185509  0.1444582704  0.107990424 -0.030500412
#> [121,] 0.06754172 -0.0182892105  0.1539850220  0.135793574 -0.027071151
#> [122,] 0.04859916 -0.0155838689  0.1332736780  0.092443071 -0.035841685
#> [123,] 0.05077267 -0.0197717208  0.1586949266  0.132510197 -0.047787212
#> [124,] 0.02847493 -0.0148588244  0.0801863593  0.080460897 -0.014633755
#> [125,] 0.03284169 -0.0131980927  0.1123161226  0.101830143 -0.033982861
#> [126,] 0.06531472 -0.0203211664  0.1703541492  0.117814780 -0.042757615
#> [127,] 0.05461326 -0.0166079372  0.1343961503  0.086333125 -0.014095737
#> [128,] 0.05331193 -0.0094484726  0.1246074493  0.101839576 -0.023943671
#> [129,] 0.04474296 -0.0113307718  0.1220700867  0.130980091 -0.046515964
#> [130,] 0.04938224 -0.0272979046  0.0676637571  0.091086602 -0.026899372
#> [131,] 0.11472602 -0.0304625887  0.1623155316  0.099993201  0.041255739
#> [132,] 0.05442077 -0.0115689215  0.1142688415  0.091884049 -0.004327310
#> [133,] 0.07257124 -0.0172428002  0.0881092549  0.066325456  0.003766827
#> [134,] 0.04346204 -0.0143067889  0.1263097455  0.102292110 -0.021747724
#> [135,] 0.05033496 -0.0205829864  0.1584434487  0.137432763 -0.023369830
#> [136,] 0.07744479  0.0177506686  0.1223575929  0.076162040 -0.040236020
#> [137,] 0.04393008 -0.0217277057  0.0964575903  0.096534804 -0.039603311
#> [138,] 0.03424794 -0.0092221856  0.0972205004  0.091990851 -0.033918697
#> [139,] 0.02788098 -0.0145096750  0.0805323041  0.081799550 -0.018629458
#> [140,] 0.04269385 -0.0169687194  0.1166209061  0.094299598 -0.045904049
#> [141,] 0.06964749 -0.0154983573  0.1643106298  0.120735377 -0.056018862
#> [142,] 0.09606941 -0.0074415351  0.1212719282  0.081176125 -0.016714625
#> [143,] 0.03611613 -0.0139515944  0.0987029929  0.060572411 -0.037840663
#> [144,] 0.04242339 -0.0185686815  0.1282847197  0.107985997 -0.027892346
#> [145,] 0.07491364 -0.0293482572  0.1409971040  0.149055892 -0.041209826
#> [146,] 0.07201938 -0.0308558356  0.1460212190  0.136025656 -0.043397177
#> [147,] 0.08813624 -0.0150112369  0.0671079000  0.003160240  0.013412656
#> [148,] 0.06767412 -0.0141228932  0.0197455467 -0.013119086  0.049870394
#> [149,] 0.04725252 -0.0122263941  0.0280846301 -0.042963579  0.016498276
#> [150,] 0.05712851 -0.0233146103 -0.0106265192 -0.015342507  0.028312222
#> [151,] 0.09369239 -0.0062721267  0.0348605360 -0.018985060  0.067761192
#> [152,] 0.12835980  0.0031709559  0.0241019009 -0.078639665  0.082749344
#> [153,] 0.13239762 -0.0003082346  0.0049827778 -0.056756511  0.082716459
#> [154,] 0.07276650 -0.0123644559  0.0171313599 -0.035674128  0.047012484
#> [155,] 0.05815623 -0.0181548358  0.0147414003 -0.041082249  0.016725725
#> [156,] 0.10240034 -0.0058468654  0.0257310203 -0.058684280  0.080334255
#> [157,] 0.06478343 -0.0178670164 -0.0043098246 -0.016318394  0.030764226
#> [158,] 0.03813960 -0.0107468437  0.0087575733 -0.024700461  0.040717537
#> [159,] 0.03669184 -0.0093823332  0.0084545724 -0.023938428  0.026649079
#> [160,] 0.05491259 -0.0178222021  0.0010990852 -0.014022226  0.046987371
#> [161,] 0.03866114 -0.0110132470  0.0101864876 -0.029517651  0.029213018
#> [162,] 0.04271390 -0.0043239374  0.0079831041 -0.021221001  0.057071532
#>                 [,6]          [,7]
#>   [1,] -0.0003161633 -0.0147973980
#>   [2,]  0.0102913629 -0.0072249757
#>   [3,] -0.0106281430 -0.0192321012
#>   [4,]  0.0103817881  0.0025755802
#>   [5,]  0.0018599033 -0.0007530618
#>   [6,]  0.0150681482  0.0246015825
#>   [7,]  0.0126900819  0.0348266017
#>   [8,]  0.0086381275 -0.0007310664
#>   [9,]  0.0093641942  0.0216417821
#>  [10,]  0.0189037436  0.0509418313
#>  [11,] -0.0011286673  0.0218479366
#>  [12,]  0.0119440064  0.0353348444
#>  [13,]  0.0259094013  0.0119124681
#>  [14,]  0.0019145582  0.0067206622
#>  [15,]  0.0112068679  0.0283712803
#>  [16,]  0.0173622938  0.0530857118
#>  [17,]  0.0075555312  0.0291821690
#>  [18,]  0.0160876767  0.0035570380
#>  [19,]  0.0151178617  0.0217990237
#>  [20,]  0.0443193450  0.0571444424
#>  [21,]  0.0285862769  0.0415508835
#>  [22,]  0.0191293099  0.0493139799
#>  [23,] -0.1380124178 -0.1283069849
#>  [24,]  0.0102008754 -0.0363641936
#>  [25,]  0.0045769044 -0.0237280956
#>  [26,]  0.0010725818 -0.0665145758
#>  [27,]  0.0112068679  0.0283712802
#>  [28,] -0.0003488678 -0.0290969039
#>  [29,]  0.0226419922  0.0562085797
#>  [30,]  0.0366512777  0.0723216076
#>  [31,] -0.0509366006 -0.0649614728
#>  [32,] -0.0145407808 -0.0049625592
#>  [33,]  0.0223005171  0.0439077344
#>  [34,]  0.0357222921  0.0836061154
#>  [35,] -0.0584661301 -0.0541668851
#>  [36,]  0.0719759900  0.0545613795
#>  [37,]  0.0508095892  0.0460471427
#>  [38,]  0.0103632369 -0.0984933040
#>  [39,] -0.0189575668  0.0038073648
#>  [40,]  0.0451810784  0.0470584193
#>  [41,]  0.0651187141  0.0964619203
#>  [42,]  0.0561698336 -0.1144410506
#>  [43,]  0.0393066015 -0.1089504109
#>  [44,]  0.0488794979 -0.0513525757
#>  [45,]  0.0656496741  0.0276955248
#>  [46,]  0.0532183510  0.0833043054
#>  [47,]  0.0678712541  0.0472163016
#>  [48,]  0.0543758550  0.0594135999
#>  [49,]  0.0714186610 -0.0288995698
#>  [50,]  0.0575679134 -0.0273191836
#>  [51,] -0.0203042458 -0.0544854899
#>  [52,]  0.0358693711 -0.0716046629
#>  [53,]  0.0321287180  0.0828383067
#>  [54,]  0.0152187668  0.0331915972
#>  [55,]  0.0454984693  0.0982177884
#>  [56,]  0.0646643672  0.0723387082
#>  [57,] -0.0006692876 -0.0986438972
#>  [58,]  0.0591171105  0.0444723539
#>  [59,]  0.0116238722  0.0685833711
#>  [60,]  0.1068981129  0.1413466986
#>  [61,]  0.0897957690 -0.0643792841
#>  [62,]  0.0390047665 -0.0825127737
#>  [63,]  0.1039486040  0.0870243636
#>  [64,]  0.1108129066  0.0644192511
#>  [65,]  0.1045400313  0.1244765204
#>  [66,]  0.0540143109  0.0847501870
#>  [67,]  0.0890897359  0.1068962120
#>  [68,]  0.1108621395  0.0658717311
#>  [69,]  0.0772773583  0.1073083584
#>  [70,]  0.1241603480  0.0609842048
#>  [71,]  0.0967972854  0.1216687133
#>  [72,]  0.0679319810  0.0090484933
#>  [73,]  0.0562377001 -0.0417906965
#>  [74,]  0.0716112530  0.0869336623
#>  [75,]  0.0466658189 -0.2447689653
#>  [76,] -0.0119113732 -0.0918778929
#>  [77,]  0.0127864805 -0.0891119120
#>  [78,]  0.0336157137 -0.1268198660
#>  [79,]  0.0771419728  0.0341251436
#>  [80,]  0.0789325167  0.0325924595
#>  [81,]  0.1181567011  0.1232904542
#>  [82,]  0.0612350172  0.0170767568
#>  [83,]  0.0367043776  0.0023067128
#>  [84,]  0.0983370831 -0.0140338926
#>  [85,]  0.0479589070 -0.0376289911
#>  [86,]  0.0300604895  0.1478645670
#>  [87,]  0.0019957624  0.1136191319
#>  [88,]  0.0530228117 -0.0338389383
#>  [89,]  0.0362986201  0.0205656892
#>  [90,]  0.0265387625  0.0161493892
#>  [91,]  0.0261726424 -0.0045500316
#>  [92,] -0.0260891185 -0.0366203802
#>  [93,] -0.0722373926 -0.0867031179
#>  [94,]  0.0181171171  0.0195658380
#>  [95,]  0.0797044585  0.1447474533
#>  [96,] -0.0522517788 -0.0342001735
#>  [97,] -0.0127486729 -0.0970627305
#>  [98,]  0.0192588588 -0.0139303725
#>  [99,]  0.0186505157  0.1153161063
#> [100,]  0.0018980044 -0.0503097573
#> [101,]  0.0406286094  0.0687818584
#> [102,] -0.0598615490  0.0010364651
#> [103,]  0.0298055335 -0.1621188247
#> [104,] -0.0710480970 -0.1827067188
#> [105,]  0.0249925099 -0.0286454809
#> [106,]  0.0392831792  0.0130126835
#> [107,]  0.0328944151  0.0734893737
#> [108,] -0.0073730775  0.0158722679
#> [109,]  0.0331901657 -0.0602092152
#> [110,]  0.0289702339  0.0208707044
#> [111,]  0.0278445885  0.0062487215
#> [112,]  0.0248232731  0.0075189506
#> [113,]  0.0043844451 -0.0426877736
#> [114,]  0.0432696302  0.0262288291
#> [115,]  0.0425445915  0.1139000467
#> [116,]  0.0182342348  0.0247095797
#> [117,]  0.0285455654  0.0703907551
#> [118,]  0.0229959728  0.0710407128
#> [119,]  0.0011910345  0.0109654664
#> [120,] -0.0294583574 -0.0365527132
#> [121,]  0.0410297916  0.0278491698
#> [122,] -0.0010536127 -0.0679482294
#> [123,]  0.0057779465  0.0256744411
#> [124,]  0.0307022293  0.0747034639
#> [125,]  0.0163051326  0.0259224067
#> [126,]  0.0227340867 -0.0163809984
#> [127,] -0.0114040140 -0.0774713517
#> [128,]  0.0071966560 -0.0691489764
#> [129,]  0.0236996551  0.0679276702
#> [130,]  0.0210028412  0.1336338146
#> [131,] -0.0582407606 -0.0720738608
#> [132,]  0.0080637500 -0.0705758676
#> [133,] -0.0394293338 -0.0820956922
#> [134,] -0.0072447828 -0.0716512812
#> [135,]  0.0363480784  0.0457659885
#> [136,] -0.0026265807 -0.1656200825
#> [137,]  0.0367528980  0.0869717692
#> [138,]  0.0104643900  0.0931634997
#> [139,]  0.0213804045  0.0629510938
#> [140,]  0.0170692917  0.0359846136
#> [141,]  0.0108371993  0.0532992984
#> [142,] -0.0320092033 -0.2019958715
#> [143,] -0.0040800718 -0.0238096706
#> [144,]  0.0233066278  0.0478770068
#> [145,]  0.0166861053  0.0287040816
#> [146,]  0.0265195470  0.0799376837
#> [147,] -0.2422656791  0.0032388310
#> [148,] -0.2340920451  0.0615556273
#> [149,] -0.1914025000  0.1624804753
#> [150,] -0.1817309986  0.2143912199
#> [151,] -0.2322600088  0.0156263090
#> [152,] -0.1799912550 -0.0781985213
#> [153,] -0.1971043577 -0.0686434535
#> [154,] -0.2371493403  0.0226995533
#> [155,] -0.1725165109  0.2011126003
#> [156,] -0.1960483170  0.0130893070
#> [157,] -0.2306799322  0.0815554461
#> [158,] -0.1969791232  0.1586071137
#> [159,] -0.1861615176  0.1292591095
#> [160,] -0.1917600418  0.1018367729
#> [161,] -0.2002183317  0.1443453072
#> [162,] -0.1988799064  0.0761857857
#> 
#> $s$iter
#> [1] 3
#> 
#> $s$mprod
#> [1] 48
#> 
#> 
#> $rank
#> [1] 7
#> 
#> $outcome_model
#> 
#> Call:
#> stats::lm(formula = y ~ Z + U + 0)
#> 
#> Coefficients:
#> Zsexfemale    Zsexmale          U1          U2          U3          U4  
#>    0.14519     0.37105     0.08570     0.08491     0.01819    -0.06179  
#>         U5          U6          U7  
#>    0.28977    -0.28659    -0.42534  
#> 
#> 
#> $mediator_model
#> 
#> Call:
#> stats::lm(formula = U ~ Z + 0)
#> 
#> Coefficients:
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]     
#> Zsexfemale   0.369365   0.062741   0.024920   0.055756  -0.041807   0.097011
#> Zsexmale     0.499845   0.067277   0.089933   0.005224   0.064549  -0.098913
#>             [,7]     
#> Zsexfemale   0.196809
#> Zsexmale    -0.140632
#> 
#> 
#> $effects
#> # A tibble: 2 × 4
#>   term         nde     nie  total
#>   <chr>      <dbl>   <dbl>  <dbl>
#> 1 Zsexfemale 0.145 -0.0896 0.0556
#> 2 Zsexmale   0.371  0.157  0.528 
#> 
#> attr(,"class")
#> [1] "network_mediation"